backend/requirements.txt
Django==6.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.2
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
djangorestframework-filters==1.0.0.dev0
faker==21.0.0
torch==2.1.2
transformers==4.35.2
scikit-learn==1.3.2
numpy==1.24.3
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
